Add function to load awards data
Add the sql statement that generates awards json from the
postgres db, as well as the python function to load the
json output to mongo. The complexity/verbosity of the sql
may be a sign that this isn't the best way to pull the info
to mongo, but using the postgres json features to craft some
bespoke json is compelling, even if this is just a short-term
solution.